% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpsrn-package.R
\docType{package}
\name{hpsrn-package}
\alias{hpsrn}
\alias{hpsrn-package}
\title{hpsrn: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChristopherNClark/hpsrn}
  \item Report bugs at \url{https://github.com/ChristopherNClark/hpsrn/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
