% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpsrn-package.R
\docType{package}
\name{hpsrn-package}
\alias{hpsrn}
\alias{hpsrn-package}
\title{hpsrn: This tests the functionality of developing an R Data Package}
\description{
This package provides each round of Household Pulse Survey data and respective codebooks in an easily accessible manner. Raw data sets of Household Pulse Survey data [Public Use Files] can be accessed at the US Census Website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChristopherNClark/hpsrn}
  \item Report bugs at \url{https://github.com/ChristopherNClark/hpsrn/issues}
}

}
\author{
\strong{Maintainer}: Prasad Bhoite \email{pbhoite@fiu.edu} (\href{https://orcid.org/0000-0003-2548-0066}{ORCID})

Authors:
\itemize{
  \item Christopher Clark \email{chrclark@fiu.edu}
  \item Krupa Patel \email{krpatel@fiu.edu}
}

}
\keyword{internal}
